+++ elasticjob-cloud/elasticjob-cloud-scheduler/src/main/java/org/apache/shardingsphere/elasticjob/cloud/scheduler/mesos/SchedulerService.java
@@ -19,22 +19,26 @@ package org.apache.shardingsphere.elasticjob.cloud.scheduler.mesos;
 
 import com.google.common.util.concurrent.Service;
 import com.netflix.fenzo.TaskScheduler;
-import java.util.Optional;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.mesos.MesosSchedulerDriver;
 import org.apache.mesos.Protos;
 import org.apache.mesos.SchedulerDriver;
 import org.apache.shardingsphere.elasticjob.cloud.console.ConsoleBootstrap;
+import org.apache.shardingsphere.elasticjob.cloud.scheduler.config.app.CloudAppConfigurationListener;
 import org.apache.shardingsphere.elasticjob.cloud.scheduler.config.job.CloudJobConfigurationListener;
 import org.apache.shardingsphere.elasticjob.cloud.scheduler.env.BootstrapEnvironment;
 import org.apache.shardingsphere.elasticjob.cloud.scheduler.env.MesosConfiguration;
 import org.apache.shardingsphere.elasticjob.cloud.scheduler.ha.FrameworkIDService;
 import org.apache.shardingsphere.elasticjob.cloud.scheduler.producer.ProducerManager;
+import org.apache.shardingsphere.elasticjob.cloud.scheduler.state.disable.app.CloudAppDisableListener;
+import org.apache.shardingsphere.elasticjob.cloud.scheduler.state.disable.job.CloudJobDisableListener;
 import org.apache.shardingsphere.elasticjob.cloud.scheduler.statistics.StatisticManager;
 import org.apache.shardingsphere.elasticjob.reg.base.CoordinatorRegistryCenter;
-import org.apache.shardingsphere.elasticjob.tracing.JobEventBus;
+import org.apache.shardingsphere.elasticjob.tracing.JobTracingEventBus;
 import org.apache.shardingsphere.elasticjob.tracing.api.TracingConfiguration;
+
+import java.util.Optional;
 
 /**
  * Scheduler service.
@@ -63,21 +67,30 @@ public final class SchedulerService {
     
     private final ReconcileService reconcileService;
     
+    private final CloudJobDisableListener cloudJobDisableListener;
+    
+    private final CloudAppConfigurationListener cloudAppConfigurationListener;
+    
+    private final CloudAppDisableListener cloudAppDisableListener;
+    
     public SchedulerService(final CoordinatorRegistryCenter regCenter) {
         env = BootstrapEnvironment.getINSTANCE();
         facadeService = new FacadeService(regCenter);
         statisticManager = StatisticManager.getInstance(regCenter, env.getTracingConfiguration().orElse(null));
         TaskScheduler taskScheduler = getTaskScheduler();
-        JobEventBus jobEventBus = getJobEventBus();
-        schedulerDriver = getSchedulerDriver(taskScheduler, jobEventBus, new FrameworkIDService(regCenter));
+        JobTracingEventBus jobTracingEventBus = getJobTracingEventBus();
+        schedulerDriver = getSchedulerDriver(taskScheduler, jobTracingEventBus, new FrameworkIDService(regCenter));
         producerManager = new ProducerManager(schedulerDriver, regCenter);
         cloudJobConfigurationListener = new CloudJobConfigurationListener(regCenter, producerManager);
-        taskLaunchScheduledService = new TaskLaunchScheduledService(schedulerDriver, taskScheduler, facadeService, jobEventBus);
+        cloudJobDisableListener = new CloudJobDisableListener(regCenter, producerManager);
+        cloudAppConfigurationListener = new CloudAppConfigurationListener(regCenter, producerManager);
+        cloudAppDisableListener = new CloudAppDisableListener(regCenter, producerManager);
+        taskLaunchScheduledService = new TaskLaunchScheduledService(schedulerDriver, taskScheduler, facadeService, jobTracingEventBus);
         reconcileService = new ReconcileService(schedulerDriver, facadeService);
         consoleBootstrap = new ConsoleBootstrap(regCenter, env.getRestfulServerConfiguration(), producerManager, reconcileService);
     }
     
-    private SchedulerDriver getSchedulerDriver(final TaskScheduler taskScheduler, final JobEventBus jobEventBus, final FrameworkIDService frameworkIDService) {
+    private SchedulerDriver getSchedulerDriver(final TaskScheduler taskScheduler, final JobTracingEventBus jobTracingEventBus, final FrameworkIDService frameworkIDService) {
         Protos.FrameworkInfo.Builder builder = Protos.FrameworkInfo.newBuilder();
         frameworkIDService.fetch().ifPresent(frameworkID -> builder.setId(Protos.FrameworkID.newBuilder().setValue(frameworkID).build()));
         Optional<String> role = env.getMesosRole();
@@ -91,7 +104,7 @@ public final class SchedulerService {
         Protos.FrameworkInfo frameworkInfo = builder.setUser(mesosConfig.getUser()).setName(frameworkName)
                 .setHostname(mesosConfig.getHostname()).setFailoverTimeout(MesosConfiguration.FRAMEWORK_FAILOVER_TIMEOUT_SECONDS)
                 .setWebuiUrl(WEB_UI_PROTOCOL + env.getFrameworkHostPort()).setCheckpoint(true).build();
-        return new MesosSchedulerDriver(new SchedulerEngine(taskScheduler, facadeService, jobEventBus, frameworkIDService, statisticManager), frameworkInfo, mesosConfig.getUrl());
+        return new MesosSchedulerDriver(new SchedulerEngine(taskScheduler, facadeService, jobTracingEventBus, frameworkIDService, statisticManager), frameworkInfo, mesosConfig.getUrl());
     }
     
     private TaskScheduler getTaskScheduler() {
@@ -103,9 +116,9 @@ public final class SchedulerService {
                 }).build();
     }
     
-    private JobEventBus getJobEventBus() {
-        Optional<TracingConfiguration> tracingConfiguration = env.getTracingConfiguration();
-        return tracingConfiguration.map(JobEventBus::new).orElseGet(JobEventBus::new);
+    private JobTracingEventBus getJobTracingEventBus() {
+        Optional<TracingConfiguration<?>> tracingConfiguration = env.getTracingConfiguration();
+        return tracingConfiguration.map(JobTracingEventBus::new).orElseGet(JobTracingEventBus::new);
     }
     
     /**
@@ -116,6 +129,9 @@ public final class SchedulerService {
         producerManager.startup();
         statisticManager.startup();
         cloudJobConfigurationListener.start();
+        cloudAppConfigurationListener.start();
+        cloudJobDisableListener.start();
+        cloudAppDisableListener.start();
         taskLaunchScheduledService.startAsync();
         consoleBootstrap.start();
         schedulerDriver.start();
@@ -131,6 +147,9 @@ public final class SchedulerService {
         consoleBootstrap.stop();
         taskLaunchScheduledService.stopAsync();
         cloudJobConfigurationListener.stop();
+        cloudAppConfigurationListener.stop();
+        cloudJobDisableListener.stop();
+        cloudAppDisableListener.stop();
         statisticManager.shutdown();
         producerManager.shutdown();
         schedulerDriver.stop(true);