+++ elasticjob-lite/elasticjob-lite-spring/elasticjob-lite-spring-boot-starter/src/main/java/org/apache/shardingsphere/elasticjob/lite/spring/boot/job/ElasticJobConfigurationProperties.java
@@ -22,48 +22,57 @@ import lombok.Setter;
 import org.apache.shardingsphere.elasticjob.api.ElasticJob;
 import org.apache.shardingsphere.elasticjob.api.JobConfiguration;
 
+import java.util.Collection;
+import java.util.LinkedList;
 import java.util.Properties;
 
+/**
+ * ElasticJob configuration properties.
+ */
 @Getter
 @Setter
-public class ElasticJobConfigurationProperties {
-
+public final class ElasticJobConfigurationProperties {
+    
     private Class<? extends ElasticJob> elasticJobClass;
-
+    
     private String elasticJobType;
-
+    
     private String cron;
-
+    
+    private String jobBootstrapBeanName;
+    
     private int shardingTotalCount;
-
+    
     private String shardingItemParameters;
-
+    
     private String jobParameter;
-
+    
     private boolean monitorExecution;
-
+    
     private boolean failover;
-
+    
     private boolean misfire;
-
+    
     private int maxTimeDiffSeconds = -1;
-
+    
     private int reconcileIntervalMinutes;
-
+    
     private String jobShardingStrategyType;
-
+    
     private String jobExecutorServiceHandlerType;
-
+    
     private String jobErrorHandlerType;
-
+    
+    private Collection<String> jobListenerTypes = new LinkedList<>();
+    
     private String description;
-
+    
     private Properties props = new Properties();
-
+    
     private boolean disabled;
-
+    
     private boolean overwrite;
-
+    
     /**
      * Convert to job configuration.
      *
@@ -76,10 +85,8 @@ public class ElasticJobConfigurationProperties {
                 .monitorExecution(monitorExecution).failover(failover).misfire(misfire)
                 .maxTimeDiffSeconds(maxTimeDiffSeconds).reconcileIntervalMinutes(reconcileIntervalMinutes)
                 .jobShardingStrategyType(jobShardingStrategyType).jobExecutorServiceHandlerType(jobExecutorServiceHandlerType).jobErrorHandlerType(jobErrorHandlerType)
-                .description(description).disabled(disabled).overwrite(overwrite).build();
-        for (Object each : props.keySet()) {
-            result.getProps().setProperty(each.toString(), props.get(each.toString()).toString());
-        }
+                .jobListenerTypes(jobListenerTypes.toArray(new String[0])).description(description).disabled(disabled).overwrite(overwrite).build();
+        props.stringPropertyNames().forEach(each -> result.getProps().setProperty(each, props.getProperty(each)));
         return result;
     }
 }