+++ elasticjob-lite/elasticjob-lite-spring/elasticjob-lite-spring-boot-starter/src/test/java/org/apache/shardingsphere/elasticjob/lite/spring/boot/job/ElasticJobSpringBootTest.java
@@ -20,7 +20,10 @@ package org.apache.shardingsphere.elasticjob.lite.spring.boot.job;
 import org.apache.shardingsphere.elasticjob.api.ElasticJob;
 import org.apache.shardingsphere.elasticjob.infra.concurrent.BlockUtils;
 import org.apache.shardingsphere.elasticjob.lite.api.bootstrap.JobBootstrap;
+import org.apache.shardingsphere.elasticjob.lite.api.bootstrap.impl.OneOffJobBootstrap;
+import org.apache.shardingsphere.elasticjob.lite.api.bootstrap.impl.ScheduleJobBootstrap;
 import org.apache.shardingsphere.elasticjob.lite.spring.boot.job.fixture.EmbedTestingServer;
+import org.apache.shardingsphere.elasticjob.lite.spring.boot.job.fixture.job.impl.CustomTestJob;
 import org.apache.shardingsphere.elasticjob.lite.spring.boot.reg.ZookeeperProperties;
 import org.apache.shardingsphere.elasticjob.reg.zookeeper.ZookeeperRegistryCenter;
 import org.apache.shardingsphere.elasticjob.tracing.api.TracingConfiguration;
@@ -33,11 +36,14 @@ import org.springframework.test.context.junit4.AbstractJUnit4SpringContextTests;
 
 import javax.sql.DataSource;
 import java.sql.SQLException;
+import java.util.Arrays;
 import java.util.Map;
 
+import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 
@@ -50,7 +56,7 @@ public class ElasticJobSpringBootTest extends AbstractJUnit4SpringContextTests {
     public static void init() {
         EmbedTestingServer.start();
     }
-
+    
     @Test
     public void assertZookeeperProperties() {
         assertNotNull(applicationContext);
@@ -58,7 +64,7 @@ public class ElasticJobSpringBootTest extends AbstractJUnit4SpringContextTests {
         assertThat(actual.getServerLists(), is(EmbedTestingServer.getConnectionString()));
         assertThat(actual.getNamespace(), is("elasticjob-lite-spring-boot-starter"));
     }
-
+    
     @Test
     public void assertRegistryCenterCreation() {
         assertNotNull(applicationContext);
@@ -67,18 +73,44 @@ public class ElasticJobSpringBootTest extends AbstractJUnit4SpringContextTests {
         zookeeperRegistryCenter.persist("/foo", "bar");
         assertThat(zookeeperRegistryCenter.get("/foo"), is("bar"));
     }
-
+    
     @Test
     public void assertTracingConfigurationCreation() throws SQLException {
         assertNotNull(applicationContext);
-        TracingConfiguration tracingConfiguration = applicationContext.getBean(TracingConfiguration.class);
-        assertNotNull(tracingConfiguration);
-        assertThat(tracingConfiguration.getType(), is("RDB"));
-        assertTrue(tracingConfiguration.getStorage() instanceof DataSource);
-        DataSource dataSource = (DataSource) tracingConfiguration.getStorage();
+        TracingConfiguration<?> tracingConfig = applicationContext.getBean(TracingConfiguration.class);
+        assertNotNull(tracingConfig);
+        assertThat(tracingConfig.getType(), is("RDB"));
+        assertTrue(tracingConfig.getStorage() instanceof DataSource);
+        DataSource dataSource = (DataSource) tracingConfig.getStorage();
         assertNotNull(dataSource.getConnection());
     }
-
+    
+    @Test
+    public void assertElasticJobProperties() {
+        assertNotNull(applicationContext);
+        ElasticJobProperties elasticJobProperties = applicationContext.getBean(ElasticJobProperties.class);
+        assertNotNull(elasticJobProperties);
+        assertNotNull(elasticJobProperties.getJobs());
+        assertThat(elasticJobProperties.getJobs().size(), is(4));
+        ElasticJobConfigurationProperties customTestJobProperties = elasticJobProperties.getJobs().get("customTestJob");
+        assertNotNull(customTestJobProperties);
+        assertThat(customTestJobProperties.getElasticJobClass(), is(CustomTestJob.class));
+        assertThat(customTestJobProperties.getJobBootstrapBeanName(), is("customTestJobBean"));
+        assertThat(customTestJobProperties.getShardingTotalCount(), is(3));
+        assertNull(customTestJobProperties.getElasticJobType());
+        assertThat(customTestJobProperties.getJobListenerTypes().size(), is(2));
+        assertThat(customTestJobProperties.getJobListenerTypes(), equalTo(Arrays.asList("NOOP", "LOG")));
+        ElasticJobConfigurationProperties printTestJobProperties = elasticJobProperties.getJobs().get("printTestJob");
+        assertNotNull(printTestJobProperties);
+        assertNull(printTestJobProperties.getElasticJobClass());
+        assertThat(printTestJobProperties.getElasticJobType(), is("PRINT"));
+        assertThat(printTestJobProperties.getJobBootstrapBeanName(), is("printTestJobBean"));
+        assertThat(printTestJobProperties.getShardingTotalCount(), is(3));
+        assertTrue(printTestJobProperties.getJobListenerTypes().isEmpty());
+        assertThat(printTestJobProperties.getProps().size(), is(1));
+        assertThat(printTestJobProperties.getProps().getProperty("print.content"), is("test print job"));
+    }
+    
     @Test
     public void assertJobScheduleCreation() {
         assertNotNull(applicationContext);
@@ -87,5 +119,26 @@ public class ElasticJobSpringBootTest extends AbstractJUnit4SpringContextTests {
         Map<String, JobBootstrap> jobBootstrapBeans = applicationContext.getBeansOfType(JobBootstrap.class);
         assertFalse(jobBootstrapBeans.isEmpty());
         BlockUtils.waitingShortTime();
+    }
+    
+    @Test
+    public void assertOneOffJobBootstrapBeanName() {
+        assertNotNull(applicationContext);
+        assertNotNull(applicationContext.getBean("customTestJobBean", OneOffJobBootstrap.class));
+        assertNotNull(applicationContext.getBean("printTestJobBean", OneOffJobBootstrap.class));
+    }
+
+    @Test
+    public void assertDefaultBeanNameWithClassJob() {
+        assertNotNull(applicationContext);
+        assertNotNull(applicationContext.getBean("defaultBeanNameClassJobScheduleJobBootstrap",
+            ScheduleJobBootstrap.class));
+    }
+
+    @Test
+    public void assertDefaultBeanNameWithTypeJob() {
+        assertNotNull(applicationContext);
+        assertNotNull(applicationContext.getBean("defaultBeanNameTypeJobScheduleJobBootstrap",
+            ScheduleJobBootstrap.class));
     }
 }